Write data sets

Write data sets to a file in the same format as read_datasets().

Usage

write_datasets(x, file = "")

Arguments

x

The data set to write. The last column must be the set number.

file

either a character string naming a file or a connection open for writing. "" indicates output to the console.

See also

Examples

x <- read_datasets(text="1 2\n3 4\n\n5 6\n7 8\n", col_names=c("obj1", "obj2"))
write_datasets(x)
#> # obj1	obj2 
#> 1 2
#> 3 4
#> 
#> 5 6
#> 7 8