Compute empirical attainment function differences

Calculate the differences between the empirical attainment functions of two data sets.

eafdiff(x, y, intervals, maximise = c(FALSE, FALSE),
  rectangles = FALSE)

Arguments

x, y
Data frames corresponding to the input data of left and right sides, respectively. Each data frame has at least three columns, the third one being the set of each point. See also read_datasets.

intervals
(`integer(1)`)
The absolute range of the differences \([0, 1]\) is partitioned into the number of intervals provided.

maximise
(logical() | logical(1))
Whether the objectives must be maximised instead of minimised. Either a single logical value that applies to all objectives or a vector of logical values, with one value per objective.

rectangles
If TRUE, the output is in the form of rectangles of the same color.

Value

With rectangle=FALSE, a data.frame containing points where there is a transition in the value of the EAF differences. The last column gives the difference in terms of sets in x minus sets in y that attain each point (i.e., negative values are differences in favour y).

Details

This function calculates the differences between the EAFs of two data sets.

See also

Examples

A1 <- read_datasets(file.path(system.file(package="eaf"), "extdata", "ALG_1_dat"))
A2 <- read_datasets(file.path(system.file(package="eaf"), "extdata", "ALG_2_dat"))
d <- eafdiff(A1, A2)
# This is large
str(d)

#>  num [1:240031, 1:3] 4.57e+09 4.58e+09 4.58e+09 4.58e+09 4.58e+09 ...

head(d)

#>            [,1]       [,2] [,3]
#> [1,] 4573361504 8344801680   -1
#> [2,] 4582996706 7969324506   -1
#> [3,] 4583258862 7969074992   -1
#> [4,] 4583298356 7968540266   -1
#> [5,] 4583376832 7967905934   -1
#> [6,] 4583412456 7967368976   -1

d <- eafdiff(A1, A2, rectangles = TRUE)
# This is large
str(d)

#>  num [1:326676, 1:5] 4.57e+09 4.58e+09 4.58e+09 4.58e+09 4.58e+09 ...

head(d)

#>            [,1]       [,2]       [,3]       [,4] [,5]
#> [1,] 4573361504 8344801680 4582996706        Inf   -1
#> [2,] 4582996706 8277735948 4632442038 8344801680   -1
#> [3,] 4582996706 8277514364 4632458126 8277735948   -1
#> [4,] 4582996706 8277462996 4632470310 8277514364   -1
#> [5,] 4582996706 8277375044 4632508156 8277462996   -1
#> [6,] 4582996706 8277237770 4632556272 8277375044   -1

# FIXME: finish this!
# Small example
X <- read_datasets(text='
1 2
2.5 1

3.5 2

2 3

4 1
')
Y <- read_datasets(text='
2.5 1

3.5 2

2 3

4 1
')