Inverted Generational Distance (IGD and IGD+) and Averaged Hausdorff Distance

Functions to compute the inverted generational distance (IGD and IGD+) and the averaged Hausdorff distance between nondominated sets of points.

igd(data, reference, maximise = FALSE)

igd_plus(data, reference, maximise = FALSE)

avg_hausdorff_dist(data, reference, maximise = FALSE, p = 1L)

Arguments

data
(matrix | data.frame)
Matrix or data frame of numerical values, where each row gives the coordinates of a point.

reference
(matrix | data.frame)
Reference set as a matrix or data.frame of numerical values.

maximise
(logical() | logical(1))
Whether the objectives must be maximised instead of minimised. Either a single logical value that applies to all objectives or a vector of logical values, with one value per objective.

p
(integer(1)) Hausdorff distance parameter (default: 1L).

Value

(numeric(1)) A single numerical value.

Details

The generational distance (GD) of a set \(A\) is defined as the distance between each point \(a \in A\) and the closest point \(r\) in a reference set \(R\), averaged over the size of \(A\). Formally,

$$GD(A,R) = \frac{1}{|A|}\left(\sum_{a\in A}\min_{r\in R} d(a,r)^p\right)^{\frac{1}{p}} $$ where: $$d(a,r) = \sqrt{\sum_{k=1}^M (a_k - r_k)^2} $$

The inverted generational distance (IGD) is calculated as \(IGD(A,R) = GD(R,A)\) with \(p=1\).

GD was first proposed by Van Veldhuizen and Lamont (1998) with \(p=2\). IGD seems to have been mentioned first by Coello Coello and Reyes-Sierra (2004) , however, some people also used the name D-metric for the same concept with \(p=1\) and later papers have often used IGD/GD with \(p=1\).

IGDX is the application of IGD to decision vectors instead of objective vectors to measure closeness and diversity in decision space (Zhou et al. 2009) . One can use the functions here directly, just passing the decision vectors as data.

The modified inverted generational distanced (IGD+) was proposed by Ishibuchi et al. (2015) to ensure that IGD+ is weakly Pareto compliant, similarly to unary epsilon. It averages over \(|R|\) within the exponent \(1/p\) and modifies the distance measure as:

$$d^+(r,a) = \sqrt{\sum_{k=1}^M (\max\{r_k - a_k, 0\})^2}$$

The average Hausdorff distance (\(\Delta_p\)) was proposed by Schütze et al. (2012) and it is calculated as:

$$\Delta_p(A,R) = \max\{ IGD_p(A,R), IGD_p(R,A) \}$$

See Bezerra et al. (2017) for a comparison.

References

Leonardo C T Bezerra, Manuel López-Ibáñez, Thomas Stützle (2017). “An Empirical Assessment of the Properties of Inverted Generational Distance Indicators on Multi- and Many-objective Optimization.” In Heike Trautmann, Günter Rudolph, Kathrin Klamroth, Oliver Schütze, Margaret M Wiecek, Yaochu Jin, Christian Grimme (eds.), Evolutionary Multi-criterion Optimization, EMO 2017, Lecture Notes in Computer Science, 31--45. Springer International Publishing, Cham, Switzerland. doi: 10.1007/978-3-319-54157-0_3 , https://doi.org/10.1007/978-3-319-54157-0_3.

Carlos A Coello Coello, Margarita Reyes-Sierra (2004). “A Study of the Parallelization of a Coevolutionary Multi-objective Evolutionary Algorithm.” In Raúl Monroy, Gustavo Arroyo-Figueroa, Luis Enrique Sucar, Humberto Sossa (eds.), Proceedings of MICAI, volume 2972 of Lecture Notes in Artificial Intelligence, 688--697. Springer, Heidelberg, Germany.

Hisao Ishibuchi, Hiroyuki Masuda, Yuki Tanigaki, Yusuke Nojima (2015). “Modified Distance Calculation in Generational Distance and Inverted Generational Distance.” In António Gaspar-Cunha, Carlos Henggeler Antunes, Carlos A Coello Coello (eds.), Evolutionary Multi-criterion Optimization, EMO 2015 Part I, volume 9018 of Lecture Notes in Computer Science, 110--125. Springer, Heidelberg, Germany.

Oliver Schütze, X Esquivel, A Lara, Carlos A Coello Coello (2012). “Using the Averaged Hausdorff Distance as a Performance Measure in Evolutionary Multiobjective Optimization.” IEEE Transactions on Evolutionary Computation, 16(4), 504--522.

David A Van Veldhuizen, Gary B Lamont (1998). “Evolutionary Computation and Convergence to a Pareto Front.” In John R Koza (ed.), Late Breaking Papers at the Genetic Programming 1998 Conference, 221--228. Van Veldhuizen and Lamont, 1998a.

A Zhou, Qingfu Zhang, Yaochu Jin (2009). “Approximating the set of Pareto-optimal solutions in both the decision and objective spaces by an estimation of distribution algorithm.” IEEE Transactions on Evolutionary Computation, 13(5), 1167--1189. doi: 10.1109/TEVC.2009.2021467 , https://doi.org/10.1109/TEVC.2009.2021467.

Author

Manuel López-Ibáñez

Examples

extdata_path <- system.file(package="eaf","extdata")
path.A1 <- file.path(extdata_path, "ALG_1_dat.xz")
path.A2 <- file.path(extdata_path, "ALG_2_dat.xz")
A1 <- read_datasets(path.A1)[,1:2]
A2 <- read_datasets(path.A2)[,1:2]
ref <- filter_dominated(rbind(A1, A2))
igd(A1, ref)

#> [1] 91888189

igd(A2, ref)

#> [1] 11351992

# IGD+ (Pareto compliant)
igd_plus(A1, ref)

#> [1] 82695357

igd_plus(A2, ref)

#> [1] 10698269

# Average Haussdorff distance
avg_hausdorff_dist(A1, ref)

#> [1] 268547627

avg_hausdorff_dist(A2, ref)

#> [1] 352613092