Compute total weighted hypervolume given a set of rectangles according to Diaz and López-Ibáñez (2020) .

TODO

whv_rect(data, rectangles, reference, maximise = FALSE)

total_whv_rect(data, rectangles, reference, maximise = FALSE,
  ideal = NULL, scalefactor = 0.1)

Arguments

data
(matrix | data.frame)
Matrix or data frame of numerical values, where each row gives the coordinates of a point.

rectangles
(matrix()) weighted rectangles that will bias the computation of the hypervolume. Maybe generated by eafdiff() with rectangles=TRUE.

reference
(numeric())
Reference point as a vector of numerical values.

maximise
(logical() | logical(1))
Whether the objectives must be maximised instead of minimised. Either a single logical value that applies to all objectives or a vector of logical values, with one value per objective.

ideal
(numeric())
Ideal point as a vector of numerical values.

scalefactor
(numeric(1)) real value within \((0,1]\) that scales the overall weight of the differences (psi \(\psi\) in the original paper).

Value

A single numerical value.

Details

TODO

References

Juan Esteban Diaz, Manuel López-Ibáñez (2020). “Incorporating Decision-Maker's Preferences into the Automatic Configuration of Bi-Objective Optimisation Algorithms.” Under review.

See also

Examples

rectangles <- as.matrix(read.table(header=FALSE, text='
 1.0  3.0  2.0  Inf    1
 2.0  3.5  2.5  Inf    2
 2.0  3.0  3.0  3.5    3
'ectangles <- as.matrix(read.table(header=FALSE, text='
 1.0  3.0  2.0  Inf    1
 2.0  3.5  2.5  Inf    2
 2.0  3.0  3.0  3.5    3
'))
whv_rect (matrix(2, ncol=2), rectangles, reference = 6)

#> [1] 4

whv_rect (matrix(c(2, 1), ncol=2), rectangles, reference = 6)

#> [1] 4

whv_rect (matrix(c(1, 2), ncol=2), rectangles, reference = 6)

#> [1] 7

total_whv_rect (matrix(2, ncol=2), rectangles, reference = 6, ideal = c(1,1))

#> [1] 26

total_whv_rect (matrix(c(2, 1), ncol=2), rectangles, reference = 6, ideal = c(1,1))

#> [1] 30

total_whv_rect (matrix(c(1, 2), ncol=2), rectangles, reference = 6, ideal = c(1,1))

#> [1] 37.5